Implement "prefix LM" feature converter.
Prefix LM is a language model where the inputs part have full visibility in the
self attention layer.

To make the feature converter for the prefix LM, the newly added function
`lm_inputs_mask` can be composed with the `decoder_feature_converter`. Please see the test for a concrete example.

PiperOrigin-RevId: 343793154