Implement feature converter for SingleStack (i.e., encoder or decoder)
SingleStack serves as both encoder and decoder depending on the argument
"autorgressive".

Use case as Decoder
A decoder-only architecture can be used in various ways. This change implements
the most basic format where the masking pattern is fully causal as well as the
Prefix LM model where the attention pattern in the inputs portion is
non-causal.

The decoder only architecture can be used as "targets" only dataset (e.g.,
standard language modeling objective) or a dataset with a notion of "inputs"
and "targets" (e.g., machine translation).

Use case as Encoder:
The inputs and targets to the encoder are aligned. Just like BERT, a sentinel
token (e.g., [CLS]) is prepended to the inputs and targets sequences. This
ensures that the model can be used for a classification task. For a packed
dataset, each sequence gets separate sentinel tokens.

PiperOrigin-RevId: 344794651