Make task and model feature dtypes "true" class constants.
Current version can lead to a subtle issue because the changes made to the
class variable `_model_feature_dtypes` when self.pack = True persists across
different object instantiations of a subclass.

To fix this issue, we can make model_feature_dtypes a constant. For additional
features associated with packing, we define another class constant
PACKING_FEATURE_DTYPES, which must be defined in the subclass as well.

In the `FeatureConverter.__call__`, we combine MODEL_FEATURE_DTYPES and
PACKING_FEATURE_DTYPES to construct the expected model feature dtypes. We use
this as reference for validation.

PiperOrigin-RevId: 346134110