Implement feature converters for encoder-decoder architecture.
Feature converters are used to move the data processing outside the modeling
code because some of the processing in mesh transformer implementation are
currently done inside the modeling code.

For a decoder trained with teacher-forcing, the input to the decoder is a
target shifted to the right by one position. Currently in the mesh_transformer
implementation, this shift is done by
mtf.transformer.transformer.autoregressive_inputs.

This CL implements a subset of the mtf functionality with tf, outside the
modeling code. It excludes the negative values used to signal where the loss
should be taken.

PiperOrigin-RevId: 346146501