Implement a feature converter for a prefix language model architecture.
Prefix LM is an architecture used when the input dataset has a notion of
"inputs" as well as "targets" (e.g., a standard supervised task). In that case,
the inputs and targets fields can be concatenated such that a decoder-only
architecture like prefix LM can be used.

PrefixLMFeatureConverter provides a binary mask, which has a value of 1
corresponding to the "inputs" part of the sequence and 0 for the "targets".
This can be used for non-causal masking of the inputs portion in the self
attention layer. For standard causal masking for the entire input sequence, the
binary mask can be simply ignored.

PrefixLMFeatureConverter includes an additional position in the attention mask.
This corresponds to the position where the EOS token of the "inputs" is a
decoder input and the first token of the "targets" is the decoder output.

The concatenation of inputs and targets are done before packing, significantly
simplifying `mtf.transformer.utils._dynamic_text2self`.

PiperOrigin-RevId: 348631061