Support non-string features in the evaluation
The model outputs a sequence of token ids. The decoded strings were returned to
the evaluator for metrics computation. This change moves the decoding to the
evaluator. Note that the decoding is a general concept. It can be decoding a
sequence of ids to a string. It can also be decoding a sequence of ids to a
transformed sequence of ids.

Users of non-string features must ensure the consistency between inputs and
outputs of decoding, postprocessors and metric_fns associated with an
evaluation task.

PiperOrigin-RevId: 349281456