Implement feature converter for encoder-only architecture such as BERT.
The inputs and targets to the encoder are aligned. Just like BERT, a sentinel
token (e.g., [CLS]) is expected to be prepended to the inputs and targets
sequences. This ensures that the model can be used for a classification task.
For a packed dataset, each sequence gets separate sentinel tokens.

PiperOrigin-RevId: 349327267