Makes the default T5 ranking evaluation metrics consistent with typical expectations for binary prediction problems.
The existing implementation of ranking evaluation metrics in T5 macro averages retrieval metrics computed over individual labels. For example, for a binary problem, the reported AUCpr is actually the average of the AUCpr for the positive class and the AUCpr for the negative class.

This behavior is unexpected for binary classification problems, whereby AUCpr is typically only reported for the positive class. This change modifies the AUCpr and AUCroc ranking evaluating metrics for two class problems to only report the metrics for the positive class.

For tasks with greater than 2 classes, the code no longer reports AUCpr or AUCroc metrics, but rather logs a warning message saying ""AUC-pr and AUC-roc are not supported when num_classes > 2"

PiperOrigin-RevId: 360245151