Adjusts preprocessor assertion to handle higher-dimensional inputs.
This allows the use of t5.data.preprocessors.select_random_chunk() even when the feature specified by feature_key is a higher-dimensional array, where we would like to slice the array (and all other features) along axis 0. This is useful for many discrete-time continuous signals, such as audio.

Any set of 1D features which passed the original assertion would still pass this updated assertion.

To avoid a possible race condition which resulted in the exception not being raised, the test for this function was modified to actually consume the resulting dataset.

PiperOrigin-RevId: 385600478