Update ReCoRD postprocessing and metrics to account for duplication of examples.
During training, record examples are duplicated, creating one example per possible label to train on. This is fine during training, but the same replication happens at validation time. At validataion time, the model prediction is compared to __any__ of the possible answers. when checking for correctness. This means that examples will multiple answers (which are replicated) get counted multiple times during evaluation. This means that getting an example with multiple answers wrong results in a larger penalty than it should, and examples with fewer answers have a smaller penalty than they should.

This change updates the record postprocessor and metric. The postprocessor now adds the the tuple `(idx/passage, idx/query)` as a grouping key to the output (which is now a dict, similar to how MultiRC is handled). The metric is updated to deduplicate based on this grouping key so we only count one (prediction, gold answers) pair during validation.

The other option to fix this would be to have split ReCoRD into multiple tasks. A training task that duplicates examples and a validation task does not. The two tasks would then be mixed together. This would have the advantage of faster validation (we don't actually process the duplicated examples) but would involve everyone who has created a ReCoRD based SeqIO task to rewrite their task. The actual change happens in the background.

This change effect validation scores only, test scores from submissions to the SuperGLUE leader board are correct.

PiperOrigin-RevId: 444408331